fix(annotation-queues): sync lockfile
