+++ packages/domain/annotation-queues/src/index.ts
@@ -2,10 +2,6 @@ export {
   ANNOTATION_QUEUE_NAME_MAX_LENGTH,
   ANNOTATION_QUEUE_SLUG_MAX_LENGTH,
   LIVE_QUEUE_DEFAULT_SAMPLING,
-  MAX_EXCERPT_LENGTH,
-  MAX_SNIPPET_EXCERPT_LENGTH,
-  MAX_STAGES_PER_PROMPT,
-  MAX_SUSPICIOUS_SNIPPETS,
   MAX_TRACES_PER_QUEUE_IMPORT,
   QUEUE_REVIEW_HOTKEYS,
   SYSTEM_QUEUE_ANNOTATOR_MAX_TOKENS,
@@ -35,29 +31,33 @@ export {
   annotationQueueItemSchema,
 } from "./entities/annotation-queue-items.ts"
 export { TooManyTracesSelectedError } from "./errors.ts"
+// Re-export strategy registry, strategies, types, and shared utilities from flagger-strategies
 export {
+  // Types
   type ConversationStage,
   type DetectionResult,
-  // Extraction helpers
+  // Shared utilities
   extractConversationStages,
-  extractJailbreakSuspiciousSnippets,
-  extractNsfwSuspiciousSnippets,
   extractUserTextMessages,
   extractWorkSignals,
+  // Strategies
   frustrationStrategy,
+  // Registry functions
   getQueueStrategy,
   getStageWorkSignals,
   hasQueueStrategy,
   jailbreakingStrategy,
   lazinessStrategy,
   listQueueStrategySlugs,
+  MAX_EXCERPT_LENGTH,
+  MAX_SNIPPET_EXCERPT_LENGTH,
+  MAX_STAGES_PER_PROMPT,
+  MAX_SUSPICIOUS_SNIPPETS,
   nsfwStrategy,
   type QueueStrategy,
-  rankStagesByLazinessLikelihood,
   rankStagesByRefusalLikelihood,
   refusalStrategy,
   type SuspiciousSnippet,
-  scoreLazinessLikelihood,
   scoreRefusalLikelihood,
   truncateExcerpt,
   type WorkSignals,
